+++ tasks/directory_tree.js
@@ -3,14 +3,13 @@
  * Released under the MIT license.
  * http://laxarjs.org/license
  */
-/*jshint node: true*/
 module.exports = function( grunt ) {
    'use strict';
 
    var fs = require( 'fs' );
    var path = require( 'path' );
    var async = require( 'async' );
-   var mktree = require( './lib/mktree' );
+   var mktree = require( '../lib/mktree' );
 
    grunt.registerMultiTask( 'directory_tree', 'Generate a JSON mapping of files ' +
       'inside a specific directory tree.', function( n ) {
@@ -19,14 +18,23 @@ module.exports = function( grunt ) {
          base: '.'
       } );
       var files = this.files;
+      var base = options.base;
       var done = this.async();
 
       async.each( files, function( file, done ) {
-         grunt.verbose.writeln( 'Directory tree: making tree of ' + file.src.length + ' files.' );
-         mktree( file.src, path.sep, function( tree ) {
-            grunt.file.write( file.dest, JSON.stringify( tree ) );
-            grunt.log.ok( 'Created directory tree mapping in "' + file.dest + '".' );
-            done();
+         var srcs = file.src.map( function( file ) {
+            return path.relative( base, file );
+         } );
+         var dest = file.dest;
+
+         grunt.verbose.writeln( 'Directory tree: making tree of ' + srcs.length + ' files.' );
+
+         mktree( base, srcs, function( err, tree ) {
+            if( !err ) {
+               grunt.file.write( dest, JSON.stringify( tree ) );
+               grunt.log.ok( 'Created directory tree mapping in "' + dest + '".' );
+            }
+            done( err );
          } );
       }, done );
    } );