+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Last Changes
 
+- [#2](https://github.com/LaxarJS/grunt-laxar/issues/2): Finally, added a lot of tests
 - [#1](https://github.com/LaxarJS/grunt-laxar/issues/1): Made lookup of Grunt tasks more resilient